+++ dev/integration_tests/android_views/ios/Flutter/Debug.xcconfig
@@ -1,0 +1,2 @@
+#include "Pods/Target Support Files/Pods-Runner/Pods-Runner.debug.xcconfig"
+#include "Generated.xcconfig"
