+++ dev/integration_tests/android_views/ios/Flutter/Release.xcconfig
@@ -1,0 +1,2 @@
+#include "Pods/Target Support Files/Pods-Runner/Pods-Runner.release.xcconfig"
+#include "Generated.xcconfig"
